business-world/automagic-otto/requirements_optimized.txt
# AutoMagic Optimized - Enhanced Dependencies
# Performance-optimized requirements for next-generation content creation

# Core Python async and performance
asyncio-tools==0.1.2
aiohttp==3.9.1
aiofiles==23.2.0
uvloop==0.19.0          # High-performance event loop (Linux/macOS)

# Configuration and environment
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# API clients with async support
openai==1.12.0
elevenlabs==2.0.0
httpx==0.24.1
aiohttp-retry==2.8.3

# Google/YouTube APIs
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0

# Media processing - optimized versions
ffmpeg-python==0.2.0
Pillow==10.2.0
imageio-ffmpeg==0.4.9

# Audio processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Video processing and optimization
opencv-python-headless==4.9.0.80
moviepy==1.0.3

# Performance and caching
redis==5.0.1           # For distributed caching
diskcache==5.6.3       # Local disk cache
cachetools==5.3.2      # Memory cache utilities
lru-cache==0.1.1       # Enhanced LRU cache

# Error handling and resilience
tenacity==8.2.3        # Advanced retry mechanisms
circuit-breaker==1.4.0 # Circuit breaker pattern
backoff==2.2.1         # Exponential backoff

# System monitoring and resources
psutil==5.9.8
py-cpuinfo==9.0.0
memory-profiler==0.61.0

# Async database (optional for metrics)
aiosqlite==0.19.0
sqlalchemy[asyncio]==2.0.25

# Web scraping for trends (optional)
aiohttp-session==2.12.0
beautifulsoup4==4.12.2
pytrends==4.9.2
selenium==4.17.2

# Scheduling and task management
schedule==1.2.1
croniter==2.0.1
celery==5.3.6          # For distributed task processing

# Logging and monitoring
structlog==23.2.0      # Structured logging
colorlog==6.8.2        # Colored console logging
prometheus-client==0.19.0  # Metrics collection

# Development and testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-benchmark==4.0.0
pytest-mock==3.12.0
pytest-cov==4.1.0
black==24.1.1          # Code formatting
isort==5.13.2          # Import sorting
mypy==1.8.0            # Type checking

# Security
cryptography==42.0.2
keyring==24.3.0        # Secure credential storage

# Utilities
click==8.1.7           # Enhanced CLI
rich==13.7.0           # Rich terminal output
tqdm==4.66.1           # Progress bars
humanize==4.9.0        # Human readable formats

# Optional ML/AI enhancements
transformers==4.37.2   # For advanced text processing
torch==2.1.2           # PyTorch for ML models
numpy==1.24.3
pandas==2.2.0          # Data analysis

# Platform specific optimizations
# Windows
pywin32==306; sys_platform == "win32"
winsound==1.0.4; sys_platform == "win32"

# Linux specific
python-magic==0.4.27; sys_platform == "linux"

# Networking and protocols
websockets==12.0       # WebSocket support
grpcio==1.60.0         # gRPC for service communication
protobuf==4.25.2

# File handling and compression
chardet==5.2.0
zipfile-deflate64==0.2.0
py7zr==0.21.0

# Data serialization
msgpack==1.0.7
orjson==3.9.13         # Fast JSON

# Optional cloud storage
boto3==1.34.34         # AWS S3
google-cloud-storage==2.10.0  # Google Cloud Storage
azure-storage-blob==12.19.0   # Azure Blob Storage

# Task queue and messaging
redis-py-cluster==2.1.3
kombu==5.3.5

# Performance profiling
py-spy==0.3.14
line-profiler==4.1.1
viztracer==0.16.1

# API rate limiting
ratelimit==2.2.1
slowapi==0.1.9

# Health checks and monitoring
healthcheck==1.3.3
watchdog==4.0.0        # File system monitoring

# Documentation (development)
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
mkdocs==1.5.3

# Build and deployment
setuptools==69.0.3
wheel==0.42.0
build==1.0.3
twine==4.0.2